refactor(cross-chain): reorganize Across adapters
