+++ apps/kyberswap-interface/src/pages/CrossChainSwap/adapters/KyberAcrossAdapter/abi.ts
@@ -1,0 +1,81 @@
+export const spokePoolPeripheryAbi = [
+  {
+    inputs: [{ internalType: 'contract IPermit2', name: '_permit2', type: 'address' }],
+    stateMutability: 'nonpayable',
+    type: 'constructor',
+  },
+  { inputs: [], name: 'InvalidMinExpectedInputAmount', type: 'error' },
+  { inputs: [], name: 'InvalidMsgValue', type: 'error' },
+  { inputs: [], name: 'InvalidNonce', type: 'error' },
+  { inputs: [], name: 'InvalidShortString', type: 'error' },
+  { inputs: [], name: 'InvalidSignature', type: 'error' },
+  { inputs: [], name: 'MinimumExpectedInputAmount', type: 'error' },
+  { inputs: [{ internalType: 'string', name: 'str', type: 'string' }], name: 'StringTooLong', type: 'error' },
+  { anonymous: false, inputs: [], name: 'EIP712DomainChanged', type: 'event' },
+  {
+    anonymous: false,
+    inputs: [
+      { indexed: false, internalType: 'address', name: 'exchange', type: 'address' },
+      { indexed: false, internalType: 'bytes', name: 'exchangeCalldata', type: 'bytes' },
+      { indexed: true, internalType: 'address', name: 'swapToken', type: 'address' },
+      { indexed: true, internalType: 'address', name: 'acrossInputToken', type: 'address' },
+      { indexed: false, internalType: 'uint256', name: 'swapTokenAmount', type: 'uint256' },
+      { indexed: false, internalType: 'uint256', name: 'acrossInputAmount', type: 'uint256' },
+      { indexed: true, internalType: 'bytes32', name: 'acrossOutputToken', type: 'bytes32' },
+      { indexed: false, internalType: 'uint256', name: 'acrossOutputAmount', type: 'uint256' },
+    ],
+    name: 'SwapBeforeBridge',
+    type: 'event',
+  },
+  {
+    inputs: [
+      {
+        components: [
+          {
+            components: [
+              { internalType: 'uint256', name: 'amount', type: 'uint256' },
+              { internalType: 'address', name: 'recipient', type: 'address' },
+            ],
+            internalType: 'struct SpokePoolPeripheryInterface.Fees',
+            name: 'submissionFees',
+            type: 'tuple',
+          },
+          {
+            components: [
+              { internalType: 'address', name: 'inputToken', type: 'address' },
+              { internalType: 'bytes32', name: 'outputToken', type: 'bytes32' },
+              { internalType: 'uint256', name: 'outputAmount', type: 'uint256' },
+              { internalType: 'address', name: 'depositor', type: 'address' },
+              { internalType: 'bytes32', name: 'recipient', type: 'bytes32' },
+              { internalType: 'uint256', name: 'destinationChainId', type: 'uint256' },
+              { internalType: 'bytes32', name: 'exclusiveRelayer', type: 'bytes32' },
+              { internalType: 'uint32', name: 'quoteTimestamp', type: 'uint32' },
+              { internalType: 'uint32', name: 'fillDeadline', type: 'uint32' },
+              { internalType: 'uint32', name: 'exclusivityParameter', type: 'uint32' },
+              { internalType: 'bytes', name: 'message', type: 'bytes' },
+            ],
+            internalType: 'struct SpokePoolPeripheryInterface.BaseDepositData',
+            name: 'depositData',
+            type: 'tuple',
+          },
+          { internalType: 'address', name: 'swapToken', type: 'address' },
+          { internalType: 'address', name: 'exchange', type: 'address' },
+          { internalType: 'enum SpokePoolPeripheryInterface.TransferType', name: 'transferType', type: 'uint8' },
+          { internalType: 'uint256', name: 'swapTokenAmount', type: 'uint256' },
+          { internalType: 'uint256', name: 'minExpectedInputTokenAmount', type: 'uint256' },
+          { internalType: 'bytes', name: 'routerCalldata', type: 'bytes' },
+          { internalType: 'bool', name: 'enableProportionalAdjustment', type: 'bool' },
+          { internalType: 'address', name: 'spokePool', type: 'address' },
+          { internalType: 'uint256', name: 'nonce', type: 'uint256' },
+        ],
+        internalType: 'struct SpokePoolPeripheryInterface.SwapAndDepositData',
+        name: 'swapAndDepositData',
+        type: 'tuple',
+      },
+    ],
+    name: 'swapAndBridge',
+    outputs: [],
+    stateMutability: 'payable',
+    type: 'function',
+  },
+] as const
